examples/postgres_checkpoint_production/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform

azure-functions
azure-functions-langgraph[postgres]>=0.6.0,<0.7.0
langgraph>=1.0,<2.0
langchain-core>=1.0,<2.0
